requirements-api.txt
fastapi==0.117.1
uvicorn==0.36.0
joblib==1.5.2
numpy==2.3.3
pandas==2.3.2
scikit-learn==1.7.2
lightgbm==4.6.0
SQLAlchemy==2.0.43
python-dotenv==1.1.1
pydantic==2.11.9
pydantic_core==2.33.2
httpx==0.28.1
psycopg2-binary==2.9.10
loguru==0.7.3
typing_extensions==4.15.0
streamlit==1.37.1
pytest==8.4.2
pytest-cov==5.0.0
